all: restructured the whole app once more
FlightLowFareSearch working fine so far

don't nag about it, it's a working progress
