+++ Dockerfile
@@ -2,8 +2,9 @@
 FROM golang:latest as builder
 
 WORKDIR /go/src/amadeus-go
-COPY . .
-COPY api /go/src/api/
+COPY amadeus-go .
+COPY api /go/src/api
+
 
 RUN go get -u github.com/golang/dep/cmd/dep
 RUN dep init && dep ensure
