+++ cmd/cli/cli.go
@@ -1,20 +1,21 @@
 package main
 
 import (
-	"Amadeus/amadeus-go/pkg/services"
-	"Amadeus/amadeus-go/pkg/transports"
+	"amadeus-go/pkg/services"
+	"amadeus-go/pkg/transports"
 
 	"context"
 	"flag"
-	"google.golang.org/grpc"
 	"log"
 	"time"
+
+	"google.golang.org/grpc"
 )
 
 func main() {
 	fs := flag.NewFlagSet("amadeus-cli", flag.ExitOnError)
 	var (
-		grpcAddr = fs.String("grpc-addr", ":8080", "gRPC listener address")
+		grpcAddr = fs.String("grpc-addr", ":8000", "gRPC listener address")
 	)
 
 	ctx, _ := context.WithTimeout(context.TODO(), time.Second*1)
